# perf: refactor error capture and stacktrace collection