--- a/CHANGELOG.asciidoc
+++ b/CHANGELOG.asciidoc
@@ -48,6 +48,9 @@ Notes:
 [float]
 ===== Bug fixes
 
+* Performance improvements have been made in error and stacktrace capture.
+  {pull}2094[#2094]
+
 
 [[release-notes-3.15.0]]
 ==== 3.15.0 - 2021/05/19