--- a/lib/agent.js
+++ b/lib/agent.js
@@ -1,21 +1,19 @@
 'use strict'
 
-var crypto = require('crypto')
 var http = require('http')
 var path = require('path')
 
-var afterAll = require('after-all-results')
 var isError = require('core-util-is').isError
 var Filters = require('object-filter-sequence')
 
 var config = require('./config')
 var connect = require('./middleware/connect')
 const constants = require('./constants')
+var errors = require('./errors')
 var Instrumentation = require('./instrumentation')
 var lambda = require('./lambda')
 var Metrics = require('./metrics')
 var parsers = require('./parsers')
-var stackman = require('./stackman')
 var symbols = require('./symbols')
 
 var IncomingMessage = http.IncomingMessage
@@ -313,153 +311,167 @@ Agent.prototype.addMetadataFilter = function (fn) {
   this._transport.addMetadataFilter(fn)
 }
 
+const EMPTY_OPTS = {}
+
+// Capture an APM server "error" event for the given `err` and send it to APM
+// server.
+//
+// Usage:
+//    captureError(err, opts, cb)
+//    captureError(err, cb)
+//
+// where:
+// - `err` is an Error instance, or a string message, or a "parameterized string
+//   message" object, e.g.:
+//      {
+//        message: "this is my message template: %d %s"},
+//        params: [ 42, "another param" ]
+//      }
+// - `opts` can include any of the following (all optional):
+//   - `opts.timestamp` - Milliseconds since the Unix epoch. Defaults to now.
+//   - `opts.user` - Object to add to `error.context.user`.
+//   - `opts.tags` - Deprecated, use `opts.labels`. Object to add to
+//     `error.context.labels`.
+//   - `opts.labels` - Object to add to `error.context.labels`.
+//   - `opts.custom` - Object to add to `error.context.custom`.
+//   - `opts.message` - if `err` is an Error instance, this string is added to
+//     `error.log.message`.
+//   - `opts.request` - HTTP request (node `IncomingMessage` instance) to use
+//     for `error.context.request`.
+//   - `opts.response` - HTTP response (node `ServerResponse` instance) to use
+//     for `error.context.response`.
+//   - `opts.handled` - Boolean indicating if this exception was handled by
+//     application code.
+//   - `opts.captureAttributes` - Boolean. Default true. Set to false to *not*
+//     include properties of `err` as attributes on the APM error event.
+//   - `opts.skipOutcome` - Boolean. Default false. Set to true to not have
+//     this captured error set `<currentSpan>.outcome = failure`.
+// - `cb` is a callback `function (captureErr, apmErrorIdString)`
 Agent.prototype.captureError = function (err, opts, cb) {
-  if (typeof opts === 'function') return this.captureError(err, null, opts)
-
-  var agent = this
-  var trans = this.currentTransaction
-  var span = this.currentSpan
-  var timestamp = normalizeTimestamp(opts && opts.timestamp)
-  var context = (span || trans || {})._context || { traceparent: {} }
-  var req = opts && opts.request instanceof IncomingMessage
-    ? opts.request
-    : trans && trans.req
-  var res = opts && opts.response instanceof ServerResponse
-    ? opts.response
-    : trans && trans.res
-  var captureAttributes = opts && opts.captureAttributes
-  const skipOutcome = opts && opts.skipOutcome
-  var _isError = isError(err)
-
-  if ((!opts || opts.handled !== false) &&
-      (agent._conf.captureErrorLogStackTraces === config.CAPTURE_ERROR_LOG_STACK_TRACES_ALWAYS ||
-       (!_isError && agent._conf.captureErrorLogStackTraces === config.CAPTURE_ERROR_LOG_STACK_TRACES_MESSAGES))
-  ) {
-    var captureLocation = {}
-    Error.captureStackTrace(captureLocation, Agent.prototype.captureError)
+  if (typeof opts === 'function') {
+    cb = opts
+    opts = EMPTY_OPTS
+  } else if (!opts) {
+    opts = EMPTY_OPTS
   }
 
-  if (!_isError) {
-    prepareError(parsers.parseMessage(err))
-  } else {
-    parsers.parseError(err, captureAttributes, agent, function (_, error) {
-      // As of now, parseError suppresses errors internally, but even if they
-      // were passed on, we would want to suppress them here anyway
-      prepareError(error)
-    })
+  const agent = this
+  let callSiteLoc = null
+  const errIsError = isError(err)
+  // XXX REVIEW-NOTE: This changes handled to be set `true` in cases where it
+  //    was sometimes left `undefined` by (I think) accident/bug of setting it
+  //    `handled = !opts || opts.handled`. The result of the latter is that
+  //    adding unrelated opts to have `captureError(err, { request: res}, cb)`
+  //    changes `handled` from true to undefined.
+  const handled = opts.handled !== false // default true
+  const shouldCaptureAttributes = opts.captureAttributes !== false // default true
+  const skipOutcome = Boolean(opts.skipOutcome)
+  const span = this.currentSpan
+  const timestampUs = (opts.timestamp
+    ? Math.floor(opts.timestamp * 1000)
+    : Date.now() * 1000)
+  const trans = this.currentTransaction
+  const traceContext = (span || trans || {})._context
+
+  // As an added feature, for *some* cases, we capture a stacktrace at the point
+  // this `captureError` was called. This is added to `error.log.stacktrace`.
+  if (handled &&
+      (agent._conf.captureErrorLogStackTraces === config.CAPTURE_ERROR_LOG_STACK_TRACES_ALWAYS ||
+       (!errIsError && agent._conf.captureErrorLogStackTraces === config.CAPTURE_ERROR_LOG_STACK_TRACES_MESSAGES))
+  ) {
+    callSiteLoc = {}
+    Error.captureStackTrace(callSiteLoc, Agent.prototype.captureError)
   }
 
   if (span && !skipOutcome) {
     span._setOutcomeFromErrorCapture(constants.OUTCOME_FAILURE)
   }
 
-  function prepareError (error) {
-    error.id = crypto.randomBytes(16).toString('hex')
-    error.parent_id = context.traceparent.id
-    error.trace_id = context.traceparent.traceId
-    error.timestamp = timestamp
-    error.context = {
+  // Move the remaining captureError processing to a later tick because:
+  // 1. This allows the calling code to continue processing. For example, for
+  //    Express instrumentation this can significantly improve latency in
+  //    the app's endpoints because the response does not proceed until the
+  //    error handlers return.
+  // 2. Gathering `error.context.response` in the same tick results in data
+  //    for a response that hasn't yet completed (no headers, unset status_code,
+  //    etc.).
+  setImmediate(() => {
+    // Gather `error.context.*`.
+    const req = (opts.request instanceof IncomingMessage
+      ? opts.request
+      : trans && trans.req)
+    const res = (opts.response instanceof ServerResponse
+      ? opts.response
+      : trans && trans.res)
+    const errorContext = {
       user: Object.assign(
         {},
         req && parsers.getUserContextFromRequest(req),
         trans && trans._user,
-        opts && opts.user
+        opts.user
       ),
       tags: Object.assign(
         {},
         trans && trans._labels,
-        opts && opts.tags,
-        opts && opts.labels
+        opts.tags,
+        opts.labels
       ),
       custom: Object.assign(
         {},
         trans && trans._custom,
-        opts && opts.custom
+        opts.custom
       )
     }
-
-    if (trans) {
-      error.transaction_id = trans.id
-      error.transaction = {
-        type: trans.type,
-        sampled: trans.sampled
-      }
-    }
-
-    if (error.exception) {
-      error.exception.handled = !opts || opts.handled
-
-      // Optional add an alternative error message as well as the exception message
-      if (opts && opts.message && opts.message !== error.exception.message && !error.log) {
-        error.log = { message: opts.message }
-      }
-    }
-
     if (req) {
-      error.context.request = parsers.getContextFromRequest(req, agent._conf, 'errors')
+      errorContext.request = parsers.getContextFromRequest(req, agent._conf, 'errors')
     }
-
     if (res) {
-      error.context.response = parsers.getContextFromResponse(res, agent._conf, true)
+      // XXX I'm curious if current tests would have caught this sync/async buglet in incomplete res.
+      errorContext.response = parsers.getContextFromResponse(res, agent._conf, true)
     }
 
-    if (captureLocation) {
-      // prepare to add a stack trace pointing to where captureError was called
-      // from. This can make it easier to debug async stack traces.
-      stackman.callsites(captureLocation, function (err, callsites) {
-        if (err) {
-          agent.logger.debug('error while getting capture location callsites: %s', err.message)
+    errors.createAPMError({
+      log: agent.logger,
+      exception: errIsError ? err : null,
+      logMessage: errIsError ? null : err,
+      shouldCaptureAttributes,
+      timestampUs,
+      handled,
+      callSiteLoc,
+      message: opts.message,
+      sourceLinesAppFrames: agent._conf.sourceLinesErrorAppFrames,
+      sourceLinesLibraryFrames: agent._conf.sourceLinesErrorLibraryFrames,
+      trans,
+      traceContext,
+      errorContext
+    }, function filterAndSendError (_err, apmError) {
+      // _err is always null from createAPMError.
+      const id = apmError.id
+
+      apmError = agent._errorFilters.process(apmError)
+      if (!apmError) {
+        agent.logger.debug('error ignored by filter %o', { id })
+        if (cb) {
+          cb(null, id)
         }
+        return
+      }
 
-        var next = afterAll(function (_, frames) {
-          // As of now, parseCallsite suppresses errors internally, but even if
-          // they were passed on, we would want to suppress them here anyway
-
-          if (frames) {
-            // In case there isn't any log object, we'll make a dummy message
-            // as the APM Server requires a message to be present if a
-            // stacktrace also present
-            if (!error.log) error.log = { message: error.exception.message }
-            error.log.stacktrace = frames
-          }
-
-          send(error)
+      if (agent._transport) {
+        agent.logger.info('Sending error to Elastic APM: %o', { id })
+        agent._transport.sendError(apmError, function () {
+          agent.flush(function (flushErr) {
+            if (cb) {
+              cb(flushErr, id)
+            }
+          })
         })
-
-        if (callsites) {
-          for (const callsite of callsites) {
-            parsers.parseCallsite(callsite, true, agent, next())
-          }
-        }
-      })
-    } else {
-      send(error)
-    }
-  }
-
-  function send (error) {
-    const id = error.id
-
-    error = agent._errorFilters.process(error)
-
-    if (!error) {
-      agent.logger.debug('error ignored by filter %o', { id })
-      if (cb) cb(null, id)
-      return
-    }
-
-    if (agent._transport) {
-      agent.logger.info('Sending error to Elastic APM: %o', { id })
-      agent._transport.sendError(error, function () {
-        agent.flush(function (err) {
-          if (cb) cb(err, id)
-        })
-      })
-    } else if (cb) {
-      // TODO: Swallow this error just as it's done in agent.flush()?
-      process.nextTick(cb.bind(null, new Error('cannot capture error before agent is started'), id))
-    }
-  }
+      } else if (cb) {
+        // TODO: Swallow this error just as it's done in agent.flush()?
+        cb(new Error('cannot capture error before agent is started'), id)
+      }
+    })
+  })
 }
 
 // The optional callback will be called with the error object after the error
@@ -498,10 +510,6 @@ Agent.prototype.flush = function (cb) {
     this.logger.warn({ err }, err.message)
     if (cb) process.nextTick(cb)
   }
-}
-
-function normalizeTimestamp (timestamp) {
-  return (timestamp > 0 && Math.floor(timestamp * 1000)) || Date.now() * 1000
 }
 
 Agent.prototype.registerMetric = function (name, labelsOrCallback, callback) {