--- a/lib/agent.js
+++ b/lib/agent.js
@@ -486,11 +486,13 @@ Agent.prototype.handleUncaughtExceptions = function (cb) {
   }
 
   this._uncaughtExceptionListener = function (err) {
+    agent.logger.debug({ err }, 'Elastic APM caught unhandled exception')
     // The stack trace of uncaught exceptions are normally written to STDERR.
     // The `uncaughtException` listener inhibits this behavor, and it's
     // therefore necessary to manually do this to not break expectations.
-    if (agent._conf.logUncaughtExceptions === true) console.error(err)
-    agent.logger.debug('Elastic APM caught unhandled exception: %s', err.message)
+    if (agent._conf.logUncaughtExceptions === true) {
+      console.error(err)
+    }
 
     agent.captureError(err, { handled: false }, function () {
       cb ? cb(err) : process.exit(1)