--- a/lib/agent.js
+++ b/lib/agent.js
@@ -334,8 +334,8 @@ const EMPTY_OPTS = {}
 //     `error.context.labels`.
 //   - `opts.labels` - Object to add to `error.context.labels`.
 //   - `opts.custom` - Object to add to `error.context.custom`.
-//   - `opts.message` - if `err` is an Error instance, this string is added to
-//     `error.log.message`.
+//   - `opts.message` - If `err` is an Error instance, this string is added to
+//     `error.log.message` (unless it matches err.message).
 //   - `opts.request` - HTTP request (node `IncomingMessage` instance) to use
 //     for `error.context.request`.
 //   - `opts.response` - HTTP response (node `ServerResponse` instance) to use