--- a/lib/agent.js
+++ b/lib/agent.js
@@ -426,7 +426,6 @@ Agent.prototype.captureError = function (err, opts, cb) {
       errorContext.request = parsers.getContextFromRequest(req, agent._conf, 'errors')
     }
     if (res) {
-      // XXX I'm curious if current tests would have caught this sync/async buglet in incomplete res.
       errorContext.response = parsers.getContextFromResponse(res, agent._conf, true)
     }