--- a/lib/agent.js
+++ b/lib/agent.js
@@ -15,6 +15,7 @@ var lambda = require('./lambda')
 var Metrics = require('./metrics')
 var parsers = require('./parsers')
 var symbols = require('./symbols')
+const { frameCacheStats } = require('./stacktraces')
 
 var IncomingMessage = http.IncomingMessage
 var ServerResponse = http.ServerResponse
@@ -80,7 +81,9 @@ Agent.prototype.destroy = function () {
 //
 // **These stats are NOT a promised interface.**
 Agent.prototype._getStats = function () {
-  const stats = {}
+  const stats = {
+    frameCache: frameCacheStats
+  }
   if (typeof this._transport._getStats === 'function') {
     stats.apmclient = this._transport._getStats()
   }