--- a/lib/agent.js
+++ b/lib/agent.js
@@ -361,11 +361,6 @@ Agent.prototype.captureError = function (err, opts, cb) {
   const agent = this
   let callSiteLoc = null
   const errIsError = isError(err)
-  // XXX REVIEW-NOTE: This changes handled to be set `true` in cases where it
-  //    was sometimes left `undefined` by (I think) accident/bug of setting it
-  //    `handled = !opts || opts.handled`. The result of the latter is that
-  //    adding unrelated opts to have `captureError(err, { request: res}, cb)`
-  //    changes `handled` from true to undefined.
   const handled = opts.handled !== false // default true
   const shouldCaptureAttributes = opts.captureAttributes !== false // default true
   const skipOutcome = Boolean(opts.skipOutcome)