--- a/lib/errors.js
+++ b/lib/errors.js
@@ -247,6 +247,9 @@ function createAPMError (args, cb) {
         }
       }
     )
+  } else {
+    numAsyncStepsRemaining++
+    setImmediate(finish)
   }
 
   function finish () {