--- a/lib/errors.js
+++ b/lib/errors.js
@@ -125,6 +125,8 @@ function attributesFromErr (err) {
 // - `args.errorContext` - An object to be included as `error.context`.
 // - `args.message` - A message string that will be included as `error.log.message`
 //   if `args.exception` is given. Ignored if `args.logMessage` is given.
+//
+// This always calls back with `cb(null, apmError)`, i.e. it doesn't fail.
 function createAPMError (args, cb) {
   let numAsyncStepsRemaining = 0 // finish() will call cb() only when this is 0.
 
@@ -261,5 +263,8 @@ function createAPMError (args, cb) {
 }
 
 module.exports = {
-  createAPMError
+  createAPMError,
+
+  // Exported for testing.
+  _moduleNameFromFrames
 }