--- a/lib/errors.js
+++ b/lib/errors.js
@@ -132,8 +132,7 @@ function createAPMError (args, cb) {
 
   const error = {
     id: crypto.randomBytes(16).toString('hex'),
-    timestamp: args.timestampUs,
-    handled: args.handled
+    timestamp: args.timestampUs
   }
   if (args.traceContext) {
     error.parent_id = args.traceContext.traceparent.id
@@ -156,7 +155,8 @@ function createAPMError (args, cb) {
     const errMsg = String(err.message)
     error.exception = {
       message: errMsg,
-      type: String(err.name)
+      type: String(err.name),
+      handled: args.handled
     }
 
     if ('code' in err) {