--- a/lib/instrumentation/modules/express.js
+++ b/lib/instrumentation/modules/express.js
@@ -62,6 +62,7 @@ module.exports = function (express, agent, { version, enabled }) {
         } else {
           handle = function (err, req, res, next) {
             if (shouldReport(err)) {
+              // XXX what about sending 'res' through instead of relying on `currentTransaction.res`?
               agent.captureError(err, { request: req })
             }
             return orig.apply(this, arguments)