--- a/lib/instrumentation/span.js
+++ b/lib/instrumentation/span.js
@@ -180,8 +180,6 @@ Span.prototype._recordStackTrace = function (obj) {
     obj = {}
     Error.captureStackTrace(obj, Span)
   }
-<<<<<<< HEAD
-=======
   this._capturedStackTrace = obj
 }