… lib/parsers.js
… lib/stackman.js
… lib/stacktraces.js
… package.json
… test/agent.js
… test/config.js
… test/errors.test.js
… test/instrumentation/modules/apollo-server-express.js
… test/instrumentation/modules/express-graphql.js
… test/parsers.js
… test/stacktraces/_mock_apm_server.js
… test/stacktraces/fixtures/Makefile
… test/stacktraces/fixtures/capture-error-string.js
… test/stacktraces/fixtures/circular-stack.js
… test/stacktraces/fixtures/dist/throw-an-error-with-sourcemap.js
… test/stacktraces/fixtures/dist/throw-an-error-with-sourcemap.js.map
… test/stacktraces/fixtures/send-a-span.js
… test/stacktraces/fixtures/src/throw-an-error-with-sourcemap.ts
… test/stacktraces/fixtures/throw-an-error.js
… test/stacktraces/fixtures/tsconfig.json
… test/stacktraces/stacktraces.test.js
… test/test.js
… test/uncaught-exceptions/log-uncaught-exceptions-off.js