Refactor agent.captureError(...) and stacktrace collection for better
performance. This moves the stackman dep into lib/stacktraces.js and
lib/errors.js to allow breaking its compatibility for perf improvements.

Refs: #2028
Fixes: #2095

### Checklist

- [x] Implement code
- [x] Tests
- [x] Add CHANGELOG.asciidoc entry