@trentm 2021-06-04
### perf test setup

- The script being tested in `node pathosapp.js err` -- a simple express app whose `/` route does the synthetic:

```
function err(req, res, next) {
    next(new Error('oof'))
}
```

(I'll propose adding this script to the repo later. I find it useful for perf testing.)

- load that up with `hey -c 50 -q 10 -z 60s http://localhost:3000/` (a 60s run of 50 concurrent callers at 10 req/s, for a total load of 500 req/s)

- Profiling with `node --prof ...` to get a flamegraph (using the 0x tool).

Ultimately the command being run here is:
```
0x -o -D "$DATADIR" -P 'hey -c 50 -q 10 -z 60s http://localhost:3000/' bin/pathosapp.js -m "$NAME" err
```

### without these changes

Testing against the current release elastic-apm-node@3.15.0

```
$ ./scripts/flame-it err-pr2 err            # the command run (my wrapper script)
CPU: ~89%                 # observed process CPU usage during the run via `htop`

# the CPU usage for the run of pathosapp.js using `process.cpuUsage(...)`
CPU usage (μs since process start):  { user: 49386192, system: 6246167 }

# the latency distribution from `hey` output
Latency distribution:
10% in 0.0015 secs
25% in 0.0028 secs
50% in 0.0047 secs
75% in 0.0077 secs
90% in 0.0110 secs
95% in 0.0124 secs
99% in 0.0168 secs
```

A screenshot of the flamegraph for what it is worth:

![Screen Shot 2021-06-04 at 4 48 14 PM](https://user-images.githubusercontent.com/46866/120872846-b33d5d00-c554-11eb-978f-33a2e36e1d36.png)

### with the current PR changes

```
$ ./scripts/flame-it err-pr10-rmgetcwd err -s

CPU: ~70%
CPU usage (μs since process start):  { user: 39872179, system: 4153932 }

Latency distribution:
  10% in 0.0011 secs
  25% in 0.0024 secs
  50% in 0.0042 secs
  75% in 0.0069 secs
  90% in 0.0092 secs
  95% in 0.0102 secs
  99% in 0.0147 secs
```

![Screen Shot 2021-06-04 at 4 51 44 PM](https://user-images.githubusercontent.com/46866/120872980-2ba41e00-c555-11eb-9b21-a5fec80fdd8d.png)


### observations

- For the same load CPU usage is down from 90% to 70%.
- For this (albeit extreme and synthetic workload: app erroring at full tilt) the latency distribution numbers are a 10-20% improvement.

(These flamegraph screenshots don't really provide much info after the fact. They are useful for focusing perf work during dev.)