@trentm 2021-06-05
## Rough summary of changes

Perf issues:
- `{error,span}.stacktrace` logic was largely in the "stackman" module.
  It had a few perf issues:
    1. It monkey-patched (or wrapped?) v8 native CallSite objects with
       a bunch of properties for convenience/sugar. Using these in
       stacktrace generation was very slow.
    2. It had a bug in its caching of sourcemap file loading for each
       stackframe: when a file had no separate sourcemap file (the common
       case), this wasn't cached. That means we would be stat'ing for
       a sourcemap file for almost every frame.
    3. A stack frame's `filename` is a relative path to cwd. The code ended
       up calling `process.cwd()` for *every* frame, which is expensive.
  Dealing with the first issue can't be done without either completely breaking
  "stackman" backwards compat, or just no longer using "stackman" and vendoring
  some of its code. I've done the latter. Error and stacktrace code has moved to
  lib/errors.js and lib/stacktraces.js respectively.
- This adds caching to the conversion of a CallSite object to an APM error
  stacktrace frame (frameFromCallSite), which had a significant impact on
  CPU usage in the perf test mentioned above (it was about 6-7% of the diff).

Other issues:
- Fix sourcemap "sourcesContent" usage (#2094).
  Test case: test/stacktraces/stacktraces.test.js