@trentm 2021-06-07
On the test failure: It failed in a single one of the test runs and failed right at the end, during Docker cleanup:

```
[2021-06-05T00:36:37.820Z] error while removing network: network docker_default id 8f9957080cae6d9f1130514145c7a06cedf877e3b26bfdb2afb04ada4c9ff3a9 has active endpoints
```