@trentm 2021-06-07
> On the test failure:

And then a re-run of tests (including the TAV and edge tests)... *worked*, but the runner lost the plot in finishing the "edge tests":

https://apm-ci.elastic.co/job/apm-agent-nodejs/job/apm-agent-nodejs-mbp/job/PR-2094/21/
Blue ocean view showing the "RC Test" step that just hung and never completed: https://apm-ci.elastic.co/blue/organizations/jenkins/apm-agent-nodejs%2Fapm-agent-nodejs-mbp/detail/PR-2094/21/pipeline/5295