@trentm 2021-06-08
@astorm FYI: I found a bug. Because (a) frame Objects cache be retrieved from a cache now and (b) because we use a JSON serializer for APM events that replaces an object it has already seen during serialization with `"[Circular]"`, if a given stacktrace has two frames at the same location, then the serializer will produce a serialized `error.exception.stacktrace` with some elements `"[Circular]"`. This happens with the Express-based pathosapp, e.g.:

```
...
                    {
                        "filename": "node_modules/express/lib/router/index.js",
                        "lineno": 284,
                        "function": "<anonymous>",
                        "library_frame": true,
                        "abs_path": "/Users/trentm/el/apm-nodejs-dev/perf/node_modules/express/lib/router/index.js",
                        "pre_context": [
                            "      }",
                            ""
                        ],
                        "context_line": "      trim_prefix(layer, layerError, layerPath, path);",
                        "post_context": [
                            "    });",
                            "  }"
                        ]
                    },
                    "[Circular]",
                    "[Circular]",
                    {
                        "filename": "node_modules/express/lib/middleware/query.js",
                        "lineno": 45,
                        "function": "query",
                        "library_frame": true,
                        "abs_path": "/Users/trentm/el/apm-nodejs-dev/perf/node_modules/express/lib/middleware/query.js",
                        "pre_context": [
                            "    }",
                            ""
                        ],
                        "context_line": "    next();",
                        "post_context": [
                            "  };",
                            "};"
                        ]
                    },
                    "[Circular]",
                    "[Circular]",
                    "[Circular]",
                    "[Circular]",
                    "[Circular]",
                    "[Circular]",
                    {
                        "filename": "node_modules/express/lib/router/index.js",
                        "lineno": 174,
                        "function": "handle",
                        "library_frame": true,
...
```

I'm working on fix.